% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_int_conversion.R
\name{hex_to_int}
\alias{hex_to_int}
\title{Convert a hex string to an integer.}
\usage{
hex_to_int(hex)
}
\arguments{
\item{hex}{A hexadecimal color value (e.g., "FF").}
}
\value{
An integer representation of the hex value.
}
\description{
Convert a hex string to an integer.
}
