% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Color_filter.R
\name{rgb_array_to_magick}
\alias{rgb_array_to_magick}
\title{Convert a 3D RGB hex array to a magick image.}
\usage{
rgb_array_to_magick(rgb_data)
}
\arguments{
\item{rgb_data}{3D RGB hex array with dimensions [3, width, height].}
}
\value{
A magick image.
}
\description{
Convert a 3D RGB hex array to a magick image.
}
