% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gray_and_thresholding.R
\name{to_greyscale}
\alias{to_greyscale}
\title{Convert RGB hex values to grayscale using the luminance formula.}
\usage{
to_greyscale(r, g, b)
}
\arguments{
\item{r, }{g, b Red, green, and blue values in hex.}
}
\value{
A vector of grayscale values in hex.
}
\description{
Convert RGB hex values to grayscale using the luminance formula.
}
